create program to add two integers
Here is the add two numbers code. For that integer value type, this code displays the right result. In this program, users are asked to enter two integers. These two integers are stored in variables. Next, these two numbers are added using the + operator, and the resulting sum is stored in the sum variable